web_app/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.12b0
bleach==4.1.0
certifi==2021.5.30
charset-normalizer==2.0.8
click==8.0.3
cssselect==1.1.0
cycler==0.10.0
Django==3.2.3
django-debug-toolbar==3.2.2
django-summernote==0.8.20.0
fake-useragent==0.1.11
importlib-metadata==4.8.2
isort==5.10.1
lxml==4.6.4
mypy-extensions==0.4.3
packaging==21.2
parse==1.19.0
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.1
pyee==8.2.2
pyparsing==2.4.7
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
text-unidecode==1.3
tomli==1.2.3
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.7
w3lib==1.22.0
webencodings==0.5.1
zipp==3.6.0
